+++ database/migrations/2019_08_02_021348_creartabla_oferta_producto.php
@@ -17,8 +17,8 @@ class CreartablaOfertaProducto extends Migration
             $table->bigIncrements('id');
             $table->integer('precio_oferta');
             $table->dateTime('fecha');
-            $table->unsignedInteger('oferta_id');
-            $table->unsignedInteger('producto_id');
+            $table->unsignedBigInteger('oferta_id');
+            $table->unsignedBigInteger('producto_id');
             $table->foreign('oferta_id')->references('id')->on('oferta')->onDelete('restrict')->onUpdate('restrict');
             $table->foreign('producto_id')->references('id')->on('producto')->onDelete('restrict')->onUpdate('restrict');
             $table->charset = 'utf8mb4';
